Added Linear Least Squares regression, to start with
Still working on getting the Cochrane-Orcutt procedure worked out.  Not
familiar with it, but this seemed to be the necessary first step.  If
nothing else, least squares is common enough to be useful.  I think I've
got the running average working correctly (without loops), based on some
research that I've done, but if you see something wrong with it/know of
a better method, let me know.  It looks like it over-smoothed the data,
to me.  not sure what that's all about