Changed most of the introduction in Lesson 0
Also changed the copyright notice, referring now to "content" only and
not "code" (for which we have MIT license on the repo head.
I've added descriptions of how to use these notebooks (either download
a Python distro or in the cloud with Wakari) and I changed the section
on Libraries.