rover planning.py
preliminary code(I got writing pseudocode and it just sort of came out.
a little more complicated than it may need to be right now; think I need
to step it back.  It also has become apparent that just detecting a wall
in front to turn is not a sufficient condition to navigate the maze.  I
know we talked about it before with just the one sensor, but it makes
the logic a lot harder.  I think we should introduce the problem with
two "sensors" (front and right side), to simplify things.  Memory of
where you've been, I think, would also work well